requirements.txt
python-telegram-bot==13.15
requests
sqlite3-binary